WriteDiaryStepperに処理を戻す
